20191214 空港测试版本

+++ HARDWARE/inc/adc.h
@@ -5,6 +5,6 @@
 
 void Adc_Init(void);
 u16  Get_Adc(u8 ch); 
-u16 Get_Adc_Average(void); 
+int Get_Adc_Average(void); 
  
 #endif
